WIP: Add ray picking process
Allow users to pick along a ray defined from the mouse picked position.
The ray direction is defined in accordance with the camera direction.
The ray origins in the camera near clipping plane and is incremented
untill an intersection is found.
This commit is tagged as WIP because the casting of the ray is not handled yet.
For now, the ray will terminate if it goes under the plane y=0.
Future commits will aim to check the ray intersection with tiles in order to
properly end the picking process.