WIP: Enhance current picking process
Casting the picked point to get the selected tile.
This commit introduces a new classe defining a tile on the map.
When picking, we get the tile that lies under the current point
on the ray. Future commits will aim to check the intersection of this
point with the hovered tile.